requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
easyocr==1.7.2
filelock==3.18.0
Flask==3.1.1
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
narwhals==1.41.0
networkx==3.5
ninja==1.11.1.4
numpy==2.2.6
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.2.1
protobuf==6.31.1
pyarrow==20.0.0
pyclipper==1.3.0.post6
pydeck==0.9.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
shapely==2.1.1
six==1.17.0
smmap==5.0.2
streamlit==1.45.1
streamlit-cropperjs==0.0.9
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.6.1
toml==0.10.2
torch==2.7.0
torchvision==0.22.0
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
Werkzeug==3.1.3
